--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/web/controller/dto/request/progress/business/ProgressApproveRequest.java
@@ -1,20 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.web.controller.dto.request.progress.business;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProgressApproveRequest {
-
-    @NotNull
-    private Long helpId;
-
-    @NotNull
-    private Boolean isApproved;
-
-}
