--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/web/controller/dto/request/progress/business/ProgressPhotoRequest.java
@@ -1,18 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.web.controller.dto.request.progress.business;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-
-@Getter
-@Builder(access = AccessLevel.PROTECTED)
-@AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class ProgressPhotoRequest {
-
-    @NotNull
-    private Long helpId;
-
-}
