--- check-in-request-MS/src/main/java/com/example/checkinrequestMS/HelpAPI/web/controller/dto/response/progress/business/ProgressChangeResponse.java
@@ -1,19 +1,0 @@
-package com.example.checkinrequestMS.HelpAPI.web.controller.dto.response.progress.business;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class ProgressChangeResponse {
-    private Long id;
-    private String message;
-
-    public static ProgressChangeResponse of(Long id, String message) {
-        return ProgressChangeResponse.builder()
-                .id(id)
-                .message(message)
-                .build();
-    }
-
-}
